set default parameters for MotorControl and MotorControl pass through in buggyMotion so that they default to stopping when not given a speed.

+++ buggyMotion.cpp
@@ -127,6 +127,6 @@ void buggyMotion::initMotion() {
 	_motionInitComplete = true;
 }
 
-void buggyMotion::drive(motorConfig::Direction LDir, motorConfig::Direction RDir, int LSpeed, int RSpeed) {
+void buggyMotion::drive(motorConfig::Direction LDir, motorConfig::Direction RDir, byte LSpeed, byte RSpeed) {
 	MotorControl(LDir, RDir, LSpeed, RSpeed);
 }
